requirements.txt
cachetools
HomeAssistant-API
discord.py
python-dotenv
levenshtein
pydantic